Fix SLN build errors
