+++ external/Microsoft.SqlTools.Hosting.UnitTests/CommonObjects.cs
