+++ external/Microsoft.SqlTools.Hosting.UnitTests/ProtocolTests/EventContextTests.cs
