+++ external/Microsoft.SqlTools.Hosting.UnitTests/ProtocolTests/MessageReaderTests.cs
