+++ external/Microsoft.SqlTools.Hosting.UnitTests/ProtocolTests/MessageTests.cs
