+++ external/Microsoft.SqlTools.Hosting.UnitTests/ProtocolTests/MessageWriterTests.cs
