+++ external/Microsoft.SqlTools.Hosting.UnitTests/ProtocolTests/RequestContextTests.cs
