+++ external/Microsoft.SqlTools.Hosting.UnitTests/TaskExtensions.cs
