+++ external/Microsoft.SqlTools.Hosting.UnitTests/UtilityTests/LoggerTests.cs
