+++ external/Microsoft.SqlTools.ResourceProvider.UnitTests/ResourceProvider/Azure/AzureAuthenticationManagerTest.cs
