+++ external/Microsoft.SqlTools.ResourceProvider.UnitTests/ResourceProvider/Azure/AzureResourceWrapperTest.cs
