+++ external/Microsoft.SqlTools.ResourceProvider.UnitTests/ResourceProvider/Azure/AzureSqlServerDiscoveryProviderTest.cs
