+++ external/Microsoft.SqlTools.ResourceProvider.UnitTests/ResourceProvider/Azure/AzureSubscriptionContextTest.cs
