+++ external/Microsoft.SqlTools.ResourceProvider.UnitTests/ResourceProvider/Azure/AzureTestContext.cs
