+++ external/Microsoft.SqlTools.ResourceProvider.UnitTests/ResourceProvider/Azure/FakeAzureServerDiscoveryProvider.cs
