+++ external/Microsoft.SqlTools.ResourceProvider.UnitTests/ResourceProvider/DependencyManagerTest.cs
