+++ external/Microsoft.SqlTools.ResourceProvider.UnitTests/ResourceProvider/ExceptionUtilTest.cs
