+++ external/Microsoft.SqlTools.ResourceProvider.UnitTests/ResourceProvider/Fakes/FakeAccountManager.cs
