+++ external/Microsoft.SqlTools.ResourceProvider.UnitTests/ResourceProvider/Fakes/FakeDataFactory.cs
