+++ external/Microsoft.SqlTools.ResourceProvider.UnitTests/ResourceProvider/Fakes/FakeDatabaseResourceManager.cs
