+++ external/Microsoft.SqlTools.ResourceProvider.UnitTests/ResourceProvider/Fakes/FakeServerDiscoveryProvider.cs
