+++ external/Microsoft.SqlTools.ResourceProvider.UnitTests/ResourceProvider/Fakes/FakeServerDiscoveryProvider2.cs
