+++ external/Microsoft.SqlTools.ResourceProvider.UnitTests/ResourceProvider/FirewallErrorParserTest.cs
