+++ external/Microsoft.SqlTools.ResourceProvider.UnitTests/ResourceProvider/ServiceManagerTest.cs
